Fix libgit2 creating too many files that are not needed Fix output libgit2.so.0 not recorded Fix bootstrap.sh deleting config every run. Now only writes a tup.config file if it doesn't already exist. Fix libgit2.so not being added in Tupfile, due to misplaced close-parenthesis. Add -Wl,--as-needed to LDFLAGS.

+++ bootstrap.sh
@@ -8,7 +8,9 @@ which lua &>/dev/null
 lua_installed=$?
 
 if [[ "$lua_installed" == "0" ]]; then
-	printf "" > build/tup.config
+	if [[ ! -s "build/tup.config" ]]; then
+		printf "" > build/tup.config
+	fi
 	lua scripts/getdeps.lua
 	[[ ! "$?" == "0" ]] && exit 1
 else
